requirements.txt
beautifulsoup4
pydantic
requests
python-telegram-bot
